refactor: separate crawling logic to Class
